Handle gameserver status changes (#39)
If a gameserver service switches from active to another state (suspended,
admin locked, ...), we cannot refresh the gameserver details, because
the service is gone. We throw a exception in this case, so the GUI can
handle that case. Further we use the old data, if the refresh of the
gameserver details does not result in a proper result, to prevent small
network problems.