String representation for services. (#60)
Cast a service to a string, and you get a string representation of that
service. This is useful if you want to work with the service type outside
of the PHP codebase like storing it in a database, add it as a metadata,
to an error report or generate other output like a JSON export.