Respect rate limits (#63)
The NitrAPI has a rate limit in place. This metadata will be stored in
the header of each response. To access that, getters can be used. If a
user reaches the rate limit, the NitrAPI refuses all requests. To handle
that, a NitrapiRateLimitException will be thrown, so the client can
handle that case properly.